removed bindings sub crate moved windows/mod.rs to windows.rs
