+++ Cargo.toml
@@ -20,4 +20,6 @@ thiserror = "1"
 [target.'cfg(target_os = "windows")'.dependencies]
 windows = "0.11.0"
 memalloc = "0.1"
-bindings = { path = "src/windows/bindings" }
+
+[target.'cfg(target_os = "windows")'.build-dependencies]
+windows = "0.11.0"
