--- src/windows/bindings/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "bindings"
-version = "0.1.0"
-authors = ["nicguzzo <nicguzzo@gmail.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[dependencies]
-windows = "0.11.0"
-
-[build-dependencies]
-windows = "0.11.0"
