--- src/windows/bindings/src/lib.rs
@@ -1,1 +1,0 @@
-windows::include_bindings!();
